Separate client types into own crate, so ledger does not need it
Removes about 50 crates of dependency from ledger
