+++ client/Cargo.toml
@@ -19,6 +19,7 @@ reqwest = { version = "0.10.4", default-features = false, features = ["blocking"
 serde = "1.0.105"
 serde_derive = "1.0.103"
 serde_json = "1.0.48"
+solana-transaction-status = { path = "../transaction-status", version = "1.1.0" }
 solana-net-utils = { path = "../net-utils", version = "1.1.0" }
 solana-sdk = { path = "../sdk", version = "1.1.0" }
 solana-vote-program = { path = "../programs/vote", version = "1.1.0" }
