+++ core/Cargo.toml
@@ -44,6 +44,7 @@ serde_json = "1.0.48"
 solana-budget-program = { path = "../programs/budget", version = "1.1.0" }
 solana-clap-utils = { path = "../clap-utils", version = "1.1.0" }
 solana-client = { path = "../client", version = "1.1.0" }
+solana-transaction-status = { path = "../transaction-status", version = "1.1.0" }
 solana-faucet = { path = "../faucet", version = "1.1.0" }
 ed25519-dalek = "=1.0.0-pre.1"
 solana-ledger = { path = "../ledger", version = "1.1.0" }
