+++ ledger/Cargo.toml
@@ -28,7 +28,7 @@ reed-solomon-erasure = { package = "solana-reed-solomon-erasure", version = "4.0
 regex = "1.3.5"
 serde = "1.0.105"
 serde_bytes = "0.11.3"
-solana-client = { path = "../client", version = "1.1.0" }
+solana-transaction-status = { path = "../transaction-status", version = "1.1.0" }
 solana-genesis-programs = { path = "../genesis-programs", version = "1.1.0" }
 solana-logger = { path = "../logger", version = "1.1.0" }
 solana-measure = { path = "../measure", version = "1.1.0" }
