+++ transaction-status/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "solana-transaction-status"
+version = "1.1.0"
+description = "Solana transaction status types"
+authors = ["Solana Maintainers <maintainers@solana.com>"]
+repository = "https://github.com/solana-labs/solana"
+homepage = "https://solana.com/"
+license = "Apache-2.0"
+edition = "2018"
+
+[dependencies]
+bincode = "1.2.1"
+bs58 = "0.3.0"
+solana-sdk = { path = "../sdk", version = "1.1.0" }
+serde = "1.0.105"
+serde_derive = "1.0.103"
